+++ src/main/java/insane96mcp/enhancedai/mixin/WitchMixin.java
@@ -1,0 +1,24 @@
+package insane96mcp.enhancedai.mixin;
+
+import net.minecraft.world.entity.ai.attributes.AttributeModifier;
+import net.minecraft.world.entity.monster.Witch;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.injection.At;
+import org.spongepowered.asm.mixin.injection.ModifyArg;
+
+import java.util.UUID;
+
+@Mixin(Witch.class)
+public class WitchMixin {
+	private static final UUID SPEED_MODIFIER_DRINKING_UUID = UUID.fromString("9629aa37-c8a0-4ef9-a99a-7ec039a5a4dd");
+	private static final AttributeModifier SPEED_MODIFIER_DRINKING = new AttributeModifier(SPEED_MODIFIER_DRINKING_UUID, "Drinking speed penalty", -0.25D, AttributeModifier.Operation.MULTIPLY_BASE);
+
+	@ModifyArg(method = "aiStep", at = @At(value = "INVOKE", target = "Lnet/minecraft/world/entity/ai/attributes/AttributeInstance;addTransientModifier(Lnet/minecraft/world/entity/ai/attributes/AttributeModifier;)V"))
+	public AttributeModifier onAddSpeedPenality(AttributeModifier attributeModifier) {
+		return SPEED_MODIFIER_DRINKING;
+	}
+	@ModifyArg(method = "aiStep", at = @At(value = "INVOKE", target = "Lnet/minecraft/world/entity/ai/attributes/AttributeInstance;removeModifier(Lnet/minecraft/world/entity/ai/attributes/AttributeModifier;)V"))
+	public AttributeModifier onRemoveSpeedPenality(AttributeModifier attributeModifier) {
+		return SPEED_MODIFIER_DRINKING;
+	}
+}